companion: check that provideTokens were loaded before logging out

+++ packages/@uppy/companion/src/server/controllers/logout.js
@@ -9,7 +9,7 @@ function logout (req, res) {
   const session = req.session
   const providerName = req.params.providerName
 
-  if (req.uppy.providerTokens[providerName]) {
+  if (req.uppy.providerTokens && req.uppy.providerTokens[providerName]) {
     delete req.uppy.providerTokens[providerName]
     tokenService.removeFromCookies(res, req.uppy.options, req.uppy.provider.authProviderName)
   }
